Testing

+++ src/main/java/com/foo/CustomerJDBCDataAccessService.java
@@ -29,7 +29,7 @@ public class CustomerJDBCDataAccessService implements CustomerDao {
     }
 
     @Override
-    public Optional<Customer> selectCustomerById(Integer id) {
+    public Optional<Customer> selectCustomerById(Long id) {
         var sql = """
                 Select id , name , email , age from customer where id = ?
                 """;
@@ -56,7 +56,7 @@ public class CustomerJDBCDataAccessService implements CustomerDao {
     }
 
     @Override
-    public void deleteCustomer(Integer id) {
+    public void deleteCustomer(Long id) {
         var sql = """
                 Delete from customer where id=?
                 """;
@@ -66,7 +66,7 @@ public class CustomerJDBCDataAccessService implements CustomerDao {
     }
 
     @Override
-    public boolean existsPersonWithId(Integer id) {
+    public boolean existsPersonWithId(Long id) {
         var sql = """
                 Select count(id) from customer  where id = ? 
                 """;
