+++ src/main/java/com/foo/dao/CustomerDao.java
@@ -8,11 +8,11 @@ import java.util.Optional;
 public interface CustomerDao {
 
     List<Customer> listAllCustomers();
-    Optional<Customer>selectCustomerById(Integer id);
+    Optional<Customer>selectCustomerById(Long id);
     void insertCustomer(Customer customer);
     boolean existsPersonWithEmail(String email);
-    void deleteCustomer(Integer id );
-    boolean existsPersonWithId(Integer id);
+    void deleteCustomer(Long id );
+    boolean existsPersonWithId(Long id);
     void updateCustomer(Customer update);
 
 }
